Combine decode_dicom_data and decode_dicom_image into one source file
